add specific windows tests, move splitNetPath to utils
