+++ storage-rpc-client.go
@@ -19,10 +19,7 @@ package main
 import (
 	"errors"
 	"io"
-	"net"
 	"path"
-	"path/filepath"
-	"runtime"
 	"strconv"
 	"strings"
 )
@@ -38,28 +35,6 @@ type networkStorage struct {
 const (
 	storageRPCPath = reservedBucket + "/storage"
 )
-
-// splits network path into its components Address and Path.
-func splitNetPath(networkPath string) (netAddr, netPath string, err error) {
-	if runtime.GOOS == "windows" {
-		if volumeName := filepath.VolumeName(networkPath); volumeName != "" {
-			return "", networkPath, nil
-		}
-	}
-	networkParts := strings.SplitN(networkPath, ":", 2)
-	switch len(networkParts) {
-	case 1:
-		return "", networkPath, nil
-	case 2:
-		if networkParts[1] == "" {
-			return "", "", &net.AddrError{Err: "missing path in network path", Addr: networkPath}
-		} else if networkParts[0] == "" {
-			return "", "", &net.AddrError{Err: "missing address in network path", Addr: networkPath}
-		}
-		return networkParts[0], networkParts[1], nil
-	}
-	return networkParts[0], networkParts[1], nil
-}
 
 // Converts rpc.ServerError to underlying error. This function is
 // written so that the storageAPI errors are consistent across network
