+++ utils_test.go
@@ -16,7 +16,41 @@
 
 package main
 
-import "testing"
+import (
+	"net"
+	"testing"
+)
+
+// Test for splitNetPath
+func TestSplitNetPath(t *testing.T) {
+	testCases := []struct {
+		networkPath string
+		netAddr     string
+		netPath     string
+		err         error
+	}{
+		{"10.1.10.1:", "", "", &net.AddrError{Err: "missing path in network path", Addr: "10.1.10.1:"}},
+		{"10.1.10.1", "", "10.1.10.1", nil},
+		{"10.1.10.1://", "10.1.10.1", "//", nil},
+		{"10.1.10.1:/disk/1", "10.1.10.1", "/disk/1", nil},
+		{"10.1.10.1:\\path\\test", "10.1.10.1", "\\path\\test", nil},
+	}
+
+	for i, test := range testCases {
+		receivedAddr, receivedPath, receivedErr := splitNetPath(test.networkPath)
+		if receivedAddr != test.netAddr {
+			t.Errorf("Test case %d: Expected: %s, Received: %s", i+1, test.netAddr, receivedAddr)
+		}
+		if receivedPath != test.netPath {
+			t.Errorf("Test case %d: Expected: %s, Received: %s", i+1, test.netPath, receivedPath)
+		}
+		if test.err != nil {
+			if receivedErr == nil || receivedErr.Error() != test.err.Error() {
+				t.Errorf("Test case %d: Expected: %v, Received: %v", i+1, test.err, receivedErr)
+			}
+		}
+	}
+}
 
 // Tests maximum object size.
 func TestMaxObjectSize(t *testing.T) {
